Centralizing CVA broke styling
